workaround for libstdc++

+++ Siv3D/src/Siv3D/Color/SivColor.cpp
@@ -9,6 +9,8 @@
 //
 //-----------------------------------------------
 
+# include <cmath>
+
 # include <Siv3D/ColorHSV.hpp>
 # include <Siv3D/IntFormatter.hpp>
 # include <Siv3D/HalfFloat.hpp>
@@ -92,7 +94,7 @@ namespace s3d
 		if (0.0f < rf)
 		{
 			int32 rExp = static_cast<int32>(std::floor(std::log2f(rf)));
-			float rMantissa = (rf / std::powf(2.0f, static_cast<float>(rExp)) - 1.0f);
+			float rMantissa = (rf / std::pow(2.0f, static_cast<float>(rExp)) - 1.0f);
 
 			rExp += 15; // バイアス調整
 
@@ -119,7 +121,7 @@ namespace s3d
 		if (0.0f < gf)
 		{
 			int32 gExp = static_cast<int32>(std::floor(std::log2f(gf)));
-			float gMantissa = (gf / std::powf(2.0f, static_cast<float>(gExp)) - 1.0f);
+			float gMantissa = (gf / std::pow(2.0f, static_cast<float>(gExp)) - 1.0f);
 
 			gExp += 15; // バイアス調整
 
@@ -143,7 +145,7 @@ namespace s3d
 		if (0.0f < bf)
 		{
 			int32 bExp = static_cast<int32>(std::floor(std::log2f(bf)));
-			float bMantissa = (bf / std::powf(2.0f, static_cast<float>(bExp)) - 1.0f);
+			float bMantissa = (bf / std::pow(2.0f, static_cast<float>(bExp)) - 1.0f);
 
 			bExp += 15; // バイアス調整
